scripts/requirements.txt
# packages required by  results_uploader.py
GitPython
iniconfig
PyGithub
